work on intake more
